requirements.txt
shioaji==1.2.5
requests==2.32.2
beautifulsoup4==4.12.3
pandas==2.2.2
celery==5.4.0
apscheduler==3.10.4
Django==5.0.6
redis==6.2.0
flower==2.0.1
python-dotenv==1.0.1
yapf==0.40.2
flake8==7.0.0
isort==5.13.2
psycopg2-binary==2.9.9
djangorestframework==3.15.1
tenacity==8.3.0
django-redis==6.0.0